Added original idea
game made different than original by making the game speed increase every time the player eats food. As the game progresses it gets harder. There will be a clear goal so as to make the player feel very close to achieving but making it difficult to achieve, making the game slightlly addicitve.